# Intern label keys for LogQL parser.